requirements.txt
langchain
faiss-cpu
google-generativeai
langchain-google-genai
fastapi
uvicorn
streamlit
pypdf
langchain_community
langchain-text-splitters
python-dotenv
sentence-transformers
transformers
torch
numpy